# Conflicts:
#	src/CorePluginAPI/Game/World/IPlayerEntity.cs
#	src/Executables/Game/PacketHandlers/TokenLoginHandler.cs

+++ src/Data/Game.Persistence/Migrations/Sqlite/SqliteGameDbContextModelSnapshot.cs
@@ -311,7 +311,7 @@ namespace QuantumCore.Game.Persistence.Migrations.Sqlite
                         .HasMaxLength(24)
                         .HasColumnType("TEXT");
 
-                    b.Property<uint>("PlayTime")
+                    b.Property<ulong>("PlayTime")
                         .HasColumnType("INTEGER");
 
                     b.Property<byte>("PlayerClass")
